requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiohttp-retry==2.9.1
aiosignal==1.4.0
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.7.14
charset-normalizer==3.4.2
Django==4.2.23
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
frozenlist==1.7.0
idna==3.10
multidict==6.6.3
pillow==11.3.0
propcache==0.3.2
PyJWT==2.9.0
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
twilio==9.6.5
typing_extensions==4.14.1
urllib3==2.5.0
yarl==1.20.1
